marta_gnn/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
torch>=2.0.0
torch-geometric>=2.4.0
scikit-learn>=1.3.0
gtfs-realtime-bindings>=1.0.0
protobuf>=4.21.0
requests>=2.31.0
pyyaml>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0
networkx>=3.1
pytest>=7.4.0
optuna>=3.0.0
